serto/sensor/chama/__init__.py
"""
This module contains the classes and functions for the flow analysis.
"""
# python imports
import argparse
import os.path
import yaml
from typing import List, Callable, Any

# third-party imports

# local imports
from .input import *
from .runner import *
from .sensors import *

def configure_subparsers(sensor_subparsers: argparse.ArgumentParser):
    """
    Configure the subparser for the sensors command.
    :param analysis_subparsers: The subparsers for the analysis command.
    :return: None
    """

    chama_sensor_placement_analysis_parser = sensor_subparsers.add_parser(
        'chama',
        help='Chama sensor placement analysis'
    )

    chama_sensor_placement_analysis_parser.add_argument(
        "config",
        help="Configuration file for Chama sensor placement analysis",
        action='store',
    )

    chama_sensor_placement_analysis_parser.add_argument(
        '-o',
        '--output',
        help='Output file path',
        required=True,
        action='store'
    )


def main(parser_args: argparse.Namespace, *args, **kwargs):
    """
    Main function for the chama sensor placement analysis module.
    :param parser_args: argparse.Namespace object containing the parsed arguments from the command line interface (CLI)
    :param args: Additional arguments to pass to the main function for the chama sensor placement analysis module.
    :param kwargs: Additional keyword arguments to pass to the main function for the chama sensor placement analysis module.
    """
    pass


def chama_decorator(func: Callable[..., Any]) -> Callable[..., Any]:
    """
    Decorator for the flow analysis module.
    :param func: The function to decorate
    :return: The decorated function
    """
    def wrapper(*args, **kwargs):
        return func(*args, **kwargs)
    return wrapper


@chama_decorator
def chama_main(*args, **kwargs):
    """
    Main function for the flow analysis module.
    :param args: Additional arguments to pass to the main function for the flow analysis module.
    :param kwargs: Additional keyword arguments to pass to the main function for the flow analysis module.
    """
    pass