pyproject.toml
# This file is the build configuration file for setuptools.
# It is used to configure the build process of a project.
# For more information, refer to the following documentation:
# https://setuptools.pypa.io/en/latest/userguide/declarative_config.html
#
# Licensed under the MIT License (the "License");
# you may not use this file except in compliance with the License.
# You may obtain a copy of the License at
#
#    https://opensource.org/licenses/MIT
#
# Unless required by applicable law or agreed to in writing, software
# distributed under the License is distributed on an "AS IS" BASIS,
# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
# See the License for the specific language governing permissions and
# limitations under the License.
#
# SPDX-License-Identifier: MIT
#
# This file is part of HydroDeep.
#
# Authors: Caleb Buahin and Anne Mikelonis
#
# HydroDeep is a deep learning framework for hydrological, hydraulic and hydrodynamic applications.
#
[build-system]
requires = ["setuptools", "wheel"] # PEP 508 specifications.
build-backend = "setuptools.build_meta"

[project]
name = "serto"
dynamic = ["version"]
description = "A Python package for optimizing SWMM models for infastructure design, operational control, and emergency response applications"
readme = {file = "docs\\README.rst", content-type = "text/restructuredtext"}
license = {file = "LICENSE"}
keywords = ["SWMM", "optimization", "water", "stormwater", "modeling", "hydrology", "hydraulics", "wastewater", "infrastructure", "design", "control", "emergency", "response"]
requires-python = ">=3.8"

authors = [
    {name = "Caleb Buahin", email = "buahin.caleb@epa.gov"},
    {name = "Anne Mikelonis", email = "mikelonis.anne@epa.gov"},
]

maintainers = [
    {name = "Caleb Buahin", email = "buahin.caleb@epa.gov"},
    {name = "Anne Mikelonis", email = "mikelonis.anne@epa.gov"},
]

dependencies = [
    "pandas>=2.2.2",
    "pyyaml>=6.0.1",
    "epaswmm",
    "mpi4py>=3.1.6",
    "shapely>=2.0.5",
    "fiona>=1.9.6",
    "pyproj>=3.6.1",
    "geopandas>=1.0.1",
    "matplotlib>=3.9.1",
    "plotly>=5.23.0",
    "numpy>=1.23.0",
    "chama>=0.3.0",
    "getmac>=0.9.5",
]

classifiers = [
      # How mature is this project? Common values are
          #   3 - Alpha
          #   4 - Beta
          #   5 - Production/Stable
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Hydrology",
    "Topic :: Scientific/Engineering :: Mathematics",
    "Topic :: Scientific/Engineering :: Physics",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: GIS",
    "Topic :: Scientific/Engineering :: Visualization",
    "Topic :: Scientific/Engineering :: Information Analysis",
    "Topic :: Software Development :: Libraries",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Utilities",
]

[tool.setuptools]
include-package-data = true
packages = ['serto']
package-data = {"" = ["pyproject.toml"]}

[project.scripts]
serto = "serto:main"

